requirements.txt
# =============================================================================
# KORGAN AI — Core Dependencies
# =============================================================================

# --- Web Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
httpx>=0.27.0
python-multipart>=0.0.12

# --- Database ---
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
psycopg2-binary>=2.9.9

# --- Redis ---
redis[hiredis]>=5.2.0

# --- Vector Database ---
chromadb-client>=0.5.0

# --- LLM ---
ollama>=0.4.0

# --- Cloud LLM ---
anthropic>=0.39.0
openai>=1.55.0

# --- Task Queue ---
celery[redis]>=5.4.0

# --- Utilities ---
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
structlog>=24.4.0
orjson>=3.10.0
tenacity>=9.0.0
croniter>=5.0.0
cryptography>=43.0.0

# --- Monitoring ---
prometheus-client>=0.21.0
psutil>=6.1.0

# --- Git ---
gitpython>=3.1.43

# --- Testing ---
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
